moved dQP to queue-proxy/
