+++ tests/go.mod
@@ -4,7 +4,7 @@ go 1.16
 
 replace (
 	XDTgRPC_stream/plotter => ../plotter
-	github.com/ease-lab/xdt/dQP => ../dQP
+	github.com/ease-lab/xdt/queue-proxy/dQP => ../queue-proxy/dQP
 	github.com/ease-lab/xdt/proto/crossXDT => ../proto/crossXDT
 	github.com/ease-lab/xdt/proto/downXDT => ../proto/downXDT
 	github.com/ease-lab/xdt/proto/upXDT => ../proto/upXDT
@@ -18,7 +18,7 @@ require (
 	XDTgRPC_stream/plotter v0.0.0-00010101000000-000000000000
 	github.com/containerd/containerd v1.5.2
 	github.com/ease-lab/vhive/utils/tracing/go v0.0.0-20210624210547-e0cd5d053491
-	github.com/ease-lab/xdt/dQP v0.0.0-00010101000000-000000000000
+	github.com/ease-lab/xdt/queue-proxy/dQP v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/xdt/queue-proxy/sQP v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/xdt/sdk/go_sdk v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/xdt/utils v0.0.0-00010101000000-000000000000
