+++ tests/unit_test.go
@@ -33,7 +33,7 @@ import (
 
 	ctrdlog "github.com/containerd/containerd/log"
 
-	"github.com/ease-lab/xdt/dQP"
+	"github.com/ease-lab/xdt/queue-proxy/dQP"
 	"github.com/ease-lab/xdt/queue-proxy/sQP"
 	"github.com/ease-lab/xdt/utils"
