+++ user-functions/dQP/dQP.go
@@ -33,7 +33,7 @@ import (
 
 	ctrdlog "github.com/containerd/containerd/log"
 	tracing "github.com/ease-lab/vhive/utils/tracing/go"
-	"github.com/ease-lab/xdt/dQP"
+	"github.com/ease-lab/xdt/queue-proxy/dQP"
 	"github.com/ease-lab/xdt/utils"
 	log "github.com/sirupsen/logrus"
 	pkgnet "knative.dev/pkg/network"
